feat(apps/pool): small prepare for steer
